Initial support for on Save override of any file with matching header. Still issues with undo manager.

+++ src/ServiceStackXCode/ServiceStackXCode/ExtendNSString.m
@@ -11,7 +11,7 @@
 
 @implementation NSString (util)
 
-- (int)indexOf:(NSString *)text {
+- (NSInteger)indexOf:(NSString *)text {
     NSRange range = [self rangeOfString:text];
     if (range.length > 0) {
         return range.location;
